common: rename db module to store
