+++ backend/crates/api-server/src/main.rs
@@ -4,7 +4,7 @@ use async_graphql::{
 };
 use async_graphql_warp::{self, GraphQLResponse};
 use clap::Parser;
-use graphix_common::{api_types as schema, db::Store};
+use graphix_common::{api_types as schema, store::Store};
 use std::{convert::Infallible, net::Ipv4Addr};
 use warp::{
     http::{self, Method},
