+++ backend/crates/common/src/lib.rs
@@ -1,12 +1,12 @@
 pub mod api_types;
 pub mod bisect;
 pub mod config;
-pub mod db;
 mod indexer;
 pub mod indexing_statuses;
 pub mod network_subgraph;
 mod prometheus_metrics;
 pub mod proofs_of_indexing;
+pub mod store;
 mod types;
 
 #[cfg(any(test, feature = "tests"))]
@@ -16,7 +16,7 @@ pub use prometheus_metrics::{PrometheusExporter, PrometheusMetrics};
 
 pub mod prelude {
     pub use super::config::*;
-    pub use super::db;
     pub use super::indexer::*;
+    pub use super::store;
     pub use super::types::*;
 }
